Move namedtuple to ext and add defaults to BaseConverter
